Fixes issue-37 and adds a pagination button.
Added a pagination for the "Patch Log" with a default of 100 entries per page.
The reason is that for long Patch Logs, there might occur very quickly a memory exhaustion from PHP, which returns just a blank page.

Also one should note that no unneccessary "new Something()" is present in a loop!